requirements.txt
Flask
pytest
pytest-cov
psycopg2-binary
